:recycle: switch to jsdom 3d imag scraping.
