+++ image_manip/generate_seed_image.js
@@ -1,7 +1,7 @@
 const axios = require('axios');
 const jimp = require('jimp');
 const fs = require('fs');
-const { getAddress } = require('./nightmare_crawl');
+const { getJSDOMAddress } = require('./3d_crawl');
 
 const TYPES = require('./types');
 const POKE_URL = 'https://assets.pokemon.com/assets/cms2/img/pokedex/full';
@@ -13,7 +13,7 @@ const saveImage = async (url, pokeNumber, type, dimension) =>
     const resizedBlackBackground = image.background(0xFFFFFFFF).rgba(false).autocrop().resize(dimension, dimension);
     resizedBlackBackground.write(`seed_data/${type}/color/${pokeNumber}.png`);
     return resizedBlackBackground.dither565().greyscale().write(`seed_data/${type}/grey/${pokeNumber}.png`);
-  });
+  }).catch(e => console.log('jimp failure, maybe figure out why'));
 
 /**
 * function that writes a json file to disk describing a pokemon
@@ -67,7 +67,9 @@ module.exports.getPokemon = async pokeNumber => {
   await saveImage(`${POKE_URL}/${paddedNumber.padStart(3, 0)}.png`, paddedNumber, 'art', 475);
 
   try {
-    const modeledImageSrc = await getAddress(paddedNumber);
+    const modeledImageSrc = await getJSDOMAddress(paddedNumber);
     await saveImage(modeledImageSrc, paddedNumber, '3d', 240);
-  } catch (e) {}
+  } catch (e) {
+    console.log('really really unhandled 3d image scraper error', e);
+  }
 };
